semantic_interpreter/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-groq>=0.0.1
groq>=0.5.0
python-dotenv>=1.0.0
langchain-text-splitters>=0.0.1
